feat: display category headers in list
